refactoring, create base class for participant-view and court-building-video
