+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/individual-journey/pages/court-building-video/court-building-video.component.html
@@ -1,6 +1,6 @@
 
 <div>
-  <h1 class="vh-hide-h1">Court building video</h1>
+  <h1 class="vh-hide-h1" i18n="@@courtBuildingVideo_h1">Court building video</h1>
 </div>
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-full">
