+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/individual-journey/pages/court-building-video/court-building-video.component.spec.ts
@@ -1,34 +1,13 @@
-import { CanCreateComponent } from '../individual-base-component/component-test-bed.spec';
 import { CourtBuildingVideoComponent } from './court-building-video.component';
-import { TestModuleMetadata } from '@angular/core/testing';
 import { Component, Input } from '@angular/core';
 import { VideoUrlService } from '../../services/video-url.service';
-import { Logger } from 'src/app/services/logger';
-import { Config } from '../../../shared/models/config';
 import { IndividualJourney } from '../../individual-journey';
-
-@Component({
-  selector: 'app-video-view',
-  template: ''
-})
-class StubVideoViewComponent {
-  @Input()
-  source: string;
-}
+import { CanCreateVideoViewBaseComponent } from '../../components/video-view-base/video-view-base.component.spec';
 
 describe('CourtBuildingVideoComponent', () => {
-  it('can be created', () => {
-    CanCreateComponent(CourtBuildingVideoComponent,
-      (configuration: TestModuleMetadata) => {
-        configuration.providers.push(
-          { provide: Logger, useValue: jasmine.createSpyObj<Logger>(['getVideoFileUrlerror']) },
-          { provide: VideoUrlService, useValue: jasmine.createSpyObj<VideoUrlService>(['getVideoFileUrl']) },
-          { provide: Config, useValue: {} }
-        );
-        configuration.declarations.push(StubVideoViewComponent);
-      }
-    );
-  });
+  it('can be created', (() => {
+    CanCreateVideoViewBaseComponent(CourtBuildingVideoComponent);
+  }));
 
   describe('functionality', () => {
     let component: CourtBuildingVideoComponent;
