+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/individual-journey/pages/participant-view/participant-view.component.html
@@ -1,3 +1,6 @@
+<div>
+  <h1 class="vh-hide-h1" i18n="@@participantView_h1">Practice video hearing</h1>
+</div>
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-full">
     <app-video-view [source]="videoSource" (loaded)="videoLoaded()"></app-video-view>
